Make migrations work in Laravel 5.8 . fix foreign key issue with old laravel version

+++ migrations/2018_09_30_142113_create_form_submissions_table.php
@@ -23,7 +23,7 @@ class CreateFormSubmissionsTable extends Migration
 
             $table->unsignedInteger('form_id');
 
-            $table->unsignedBigInteger('user_id')->nullable();
+            $table->unsignedInteger('user_id')->nullable();
 
             $table->text('content');
